v1.00 relocate files, header done (before .map the merch items)
