+++ package.json
@@ -22,10 +22,10 @@
     "classnames": "^2.5.1",
     "framer-motion": "^11.11.1",
     "lottie-react": "^2.4.0",
-    "next": "14.2.14",
-    "react": "^18",
+    "next": "^14.2.23",
+    "react": "^18.3.1",
     "react-countup": "^6.5.3",
-    "react-dom": "^18",
+    "react-dom": "^18.3.1",
     "react-intersection-observer": "^9.13.1",
     "react-lottie": "^1.2.4",
     "react-multi-carousel": "^2.8.5",
