+++ src/pages/merch/MerchSection/MerchData.ts
@@ -1,6 +1,3 @@
-import BasicCrewneck from "../../images/merch/BasicCrewneck.jpg";
-import BasicCrewneckWithSleevePrint from "../../images/merch/BasicCrewneckWithSleevePrint.jpg";
-
 export type MerchProperties = {
   id: number;
   item: string;
@@ -16,7 +13,7 @@ export const merchList: MerchProperties[] = [
     colors: ["Black"],
     description: "placeholder descripton",
     id: 0,
-    image: BasicCrewneck.src,
+    image: "/image/merch/BasicCrewneck.jpg",
     item: "Original Basic Crewneck",
     price: "$29.95 CAD",
     sizes: ["S", "M", "L", "XL"],
@@ -25,7 +22,7 @@ export const merchList: MerchProperties[] = [
     colors: ["Black"],
     description: "placeholder descripton",
     id: 1,
-    image: BasicCrewneckWithSleevePrint.src,
+    image: "/image/merch/BasicCrewneckWithSleevePrint.jpg",
     item: "Crewneck with Sleeve Print",
     price: "$49.99 CAD",
     sizes: ["S", "M", "L", "XL"],
