+++ src/pages/merch/index.tsx
@@ -1,0 +1,5 @@
+import MerchPage from "./MerchSection/MerchPage";
+
+export default function MerchPagePage() {
+  return <MerchPage />;
+}
